[Week2 AS] 심화과제 완료!
